CEDS-3537 Change order of summary page rows
