+++ app/views/declaration/summary/amend_summary_page.scala.html
@@ -35,7 +35,6 @@
         locations_section: locations_section,
         transaction_section: transaction_section,
         items_section: items_section,
-        warehouse_section: warehouse_section,
         transport_section: transport_section,
         label: label,
         link: link
@@ -62,8 +61,6 @@
  @transaction_section(Mode.ChangeAmend, request.cacheModel)
 
  @items_section(Mode.ChangeAmend, request.cacheModel)
-
- @warehouse_section(Mode.ChangeAmend, request.cacheModel)
 
  @transport_section(Mode.ChangeAmend, request.cacheModel)
