--- app/views/declaration/summary/warehouse_section.scala.html
@@ -1,101 +1,0 @@
-@*
- * Copyright 2021 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import models.declaration.Locations
-@import models.Mode
-@import uk.gov.hmrc.govukfrontend.views.html.components._
-@import views.html.components.gds.linkContent
-@import views.html.components.gds.summary_list
-@import views.components.gds.ActionItemBuilder._
-
-@this(summaryList: summary_list, linkContent: linkContent)
-
-@(mode: Mode, declarationData: ExportsDeclaration, actionsEnabled: Boolean = true)(implicit messages: Messages)
-
-@sectionRequired(locations: Locations) = @{
-    locations.warehouseIdentification.isDefined || locations.supervisingCustomsOffice.isDefined || locations.inlandModeOfTransportCode
-}
-
-@actions(action: Actions) = @{
-    if(actionsEnabled) Some(action)
-    else None
-}
-
-@if(sectionRequired(declarationData.locations)) {
-    @summaryList("declaration-warehouse-summary", Some(messages("declaration.summary.warehouse")), Seq(
-        declarationData.locations.warehouseIdentification.map(warehouseIdentification =>
-            SummaryListRow(
-                classes = "warehouse-id-row",
-                key = Key(
-                    content = Text(messages( if(warehouseIdentification.identificationNumber.isDefined) "declaration.summary.warehouse.id" else "declaration.summary.warehouse.no.label"))
-                ),
-                value = Value(
-                    content = Text(warehouseIdentification.identificationNumber.getOrElse(messages("site.no")))
-                ),
-                actions = actions(Actions(
-                    items = Seq(
-                        actionItem(
-                            href = controllers.declaration.routes.WarehouseIdentificationController.displayPage(mode).url,
-                            content = HtmlContent(linkContent(messages("site.change"))),
-                            visuallyHiddenText = Some(messages("declaration.summary.warehouse.id.change"))
-                        )
-                    )
-                ))
-            )
-        ),
-        declarationData.locations.supervisingCustomsOffice.map(supervisingCustomsOffice =>
-            SummaryListRow(
-                classes = "supervising-office-row",
-                key = Key(
-                    content = Text(messages("declaration.summary.warehouse.supervisingOffice"))
-                ),
-                value = Value(
-                    content = Text(supervisingCustomsOffice.supervisingCustomsOffice.getOrElse(""))
-                ),
-                actions = actions(Actions(
-                    items = Seq(
-                        actionItem(
-                            href = controllers.declaration.routes.SupervisingCustomsOfficeController.displayPage(mode).url,
-                            content = HtmlContent(linkContent(messages("site.change"))),
-                            visuallyHiddenText = Some(messages("declaration.summary.warehouse.supervisingOffice.change"))
-                        )
-                    )
-                ))
-            )
-        ),
-        declarationData.locations.inlandModeOfTransportCode.map(inlandModeOfTransportCode =>
-            SummaryListRow(
-                classes = "mode-of-transport-row",
-                key = Key(
-                    content = Text(messages("declaration.summary.warehouse.inlandModeOfTransport"))
-                ),
-                value = Value(
-                    content = Text(inlandModeOfTransportCode.inlandModeOfTransportCode.map(code => messages(s"declaration.summary.warehouse.inlandModeOfTransport.$code")).getOrElse(""))
-                ),
-                actions = actions(Actions(
-                    items = Seq(
-                        actionItem(
-                            href = controllers.declaration.routes.InlandTransportDetailsController.displayPage(mode).url,
-                            content = HtmlContent(linkContent(messages("site.change"))),
-                            visuallyHiddenText = Some(messages("declaration.summary.warehouse.inlandModeOfTransport.change"))
-                        )
-                    )
-                ))
-            )
-        )
-    )
-    )
-}
